refactor(NcHeaderMenu): export as Typescript module and reorder script tags
