+++ src/components/index.ts
@@ -56,7 +56,7 @@ export { default as NcEmojiPicker } from './NcEmojiPicker/index.js'
 export { default as NcEmptyContent } from './NcEmptyContent/index.ts'
 export { default as NcGuestContent } from './NcGuestContent/index.ts'
 export { default as NcHeaderButton } from './NcHeaderButton/index.ts'
-export { default as NcHeaderMenu } from './NcHeaderMenu/index.js'
+export { default as NcHeaderMenu } from './NcHeaderMenu/index.ts'
 export { default as NcHighlight } from './NcHighlight/index.js'
 export { default as NcIconSvgWrapper } from './NcIconSvgWrapper/index.ts'
 export { default as NcInputField } from './NcInputField/index.ts'
